requirements.txt
streamlit
PyMuPDF
langchain
google-generativeai
python-dotenv
